+++ src/main/kotlin/no/nav/bidrag/transport/behandling/grunnlag/reponse/OppdaterGrunnlagspakkeDto.kt
@@ -1,0 +1,31 @@
+package no.nav.bidrag.transport.behandling.grunnlag.reponse
+
+import io.swagger.v3.oas.annotations.media.Schema
+import no.nav.bidrag.domain.enums.GrunnlagRequestType
+import no.nav.bidrag.domain.enums.GrunnlagsRequestStatus
+
+@Schema(description = "Respons ved oppdatering av  grunnlagspakke")
+data class OppdaterGrunnlagspakkeDto(
+
+    @Schema(description = "Grunnlagspakke-id")
+    val grunnlagspakkeId: Int,
+
+    @Schema(description = "Liste over grunnlagene som er hentet inn med person-id og status")
+    val grunnlagTypeResponsListe: List<OppdaterGrunnlagDto>
+)
+
+data class OppdaterGrunnlagDto(
+
+    @Schema(description = "Hvilken type grunnlag som er hentet")
+    val type: GrunnlagRequestType,
+
+    @Schema(description = "Angir personId som grunnlag er hentet for")
+    val personId: String,
+
+    @Schema(description = "Status for utført kall")
+    val status: GrunnlagsRequestStatus,
+
+    @Schema(description = "Statusmelding for utført kall")
+    val statusMelding: String
+
+)
