+++ src/main/kotlin/no/nav/bidrag/transport/behandling/grunnlag/request/OppdaterGrunnlagspakkeRequestDto.kt
@@ -2,23 +2,10 @@ package no.nav.bidrag.transport.behandling.grunnlag.request
 
 import io.swagger.v3.oas.annotations.media.Schema
 import jakarta.validation.Valid
-import jakarta.validation.constraints.NotBlank
 import jakarta.validation.constraints.NotEmpty
 import jakarta.validation.constraints.Pattern
-import no.nav.bidrag.behandling.felles.enums.Formaal
-import no.nav.bidrag.behandling.felles.enums.GrunnlagRequestType
+import no.nav.bidrag.domain.enums.GrunnlagRequestType
 import java.time.LocalDate
-
-@Schema(description = "Request for å opprette ny grunnlagspakke, uten annet innhold")
-data class OpprettGrunnlagspakkeRequestDto(
-
-    @Schema(description = "Til hvilket formål skal grunnlagspakken benyttes. BIDRAG, FORSKUDD eller SAERTILSKUDD")
-    val formaal: Formaal,
-
-    @Schema(description = "opprettet av")
-    @field:NotBlank(message = "Kan ikke være null eller blank.")
-    val opprettetAv: String
-)
 
 data class OppdaterGrunnlagspakkeRequestDto(
