+++ src/main/kotlin/no/nav/bidrag/transport/behandling/stonad/reponse/StonadDto.kt
@@ -1,14 +1,15 @@
-package no.nav.bidrag.transport.behandling.stonad.request
+package no.nav.bidrag.transport.behandling.stonad.reponse
 
 import io.swagger.v3.oas.annotations.media.Schema
-import no.nav.bidrag.behandling.felles.enums.Innkreving
-import no.nav.bidrag.behandling.felles.enums.StonadType
+import no.nav.bidrag.domain.enums.Innkreving
+import no.nav.bidrag.domain.enums.StonadType
 import java.math.BigDecimal
 import java.time.LocalDate
 import java.time.LocalDateTime
 
-@Schema(description = "Egenskaper ved en stønadsendring")
-data class OpprettStonadRequestDto(
+data class StonadDto(
+    @Schema(description = "StønadId")
+    val stonadId: Int,
 
     @Schema(description = "Stønadstype")
     val type: StonadType,
@@ -34,18 +35,32 @@ data class OpprettStonadRequestDto(
     @Schema(description = "opprettet_av")
     val opprettetAv: String,
 
+    @Schema(description = "opprettet tidspunkt")
+    val opprettetTidspunkt: LocalDateTime,
+
+    @Schema(description = "endret av")
+    val endretAv: String?,
+
+    @Schema(description = "når sist endret tidspunkt")
+    val endretTidspunkt: LocalDateTime?,
+
     @Schema(description = "Liste over alle perioder som inngår i stønaden")
-    val periodeListe: List<OpprettStonadPeriodeRequestDto>
+    val periodeListe: List<StonadPeriodeDto>
 )
 
-@Schema(description = "Egenskaper ved en periode")
-data class OpprettStonadPeriodeRequestDto(
+data class StonadPeriodeDto(
+
+    @Schema(description = "Periode-id")
+    val periodeId: Int,
 
     @Schema(description = "Periode fra-og-med-dato")
     val periodeFom: LocalDate,
 
     @Schema(description = "Periode til-dato")
     val periodeTil: LocalDate?,
+
+    @Schema(description = "Stonad-id")
+    val stonadId: Int,
 
     @Schema(description = "Vedtak-id")
     val vedtakId: Int,
@@ -56,7 +71,7 @@ data class OpprettStonadPeriodeRequestDto(
     @Schema(description = "Angir tidspunkt perioden eventuelt er ugyldig fra (tidspunkt for vedtak med periode som erstattet denne)")
     val gyldigTil: LocalDateTime?,
 
-    @Schema(description = "Periode gjort ugyldig av vedtak-id")
+    @Schema(description = "Periode-gjort-ugyldig-av-vedtak-id")
     val periodeGjortUgyldigAvVedtakId: Int?,
 
     @Schema(description = "Beregnet stønadsbeløp")
@@ -65,40 +80,6 @@ data class OpprettStonadPeriodeRequestDto(
     @Schema(description = "Valutakoden tilhørende stønadsbeløpet")
     val valutakode: String?,
 
-    @Schema(description = "Resultatkoden tilhørende stønadsbeløpet")
+    @Schema(description = "Resultatkode for stønaden")
     val resultatkode: String
-
-)
-
-@Schema(description = "Request for å hente stønad som matcher angitte parametre")
-data class HentStonadRequest(
-    @Schema(description = "Stønadstype")
-    val type: StonadType,
-
-    @Schema(description = "Referanse til sak")
-    val sakId: String,
-
-    @Schema(description = "Id til den som skal betale stønadem")
-    val skyldnerId: String,
-
-    @Schema(description = "Id til den som krever stønadem")
-    val kravhaverId: String
-)
-
-@Schema(description = "Request for å hente stønad og perioder som var gyldige på angitt tidspunkt")
-data class HentStonadHistoriskRequest(
-    @Schema(description = "Stønadstype")
-    val type: StonadType,
-
-    @Schema(description = "Referanse til sak")
-    val sakId: String,
-
-    @Schema(description = "Id til den som skal betale stønadem")
-    val skyldnerId: String,
-
-    @Schema(description = "Id til den som krever stønadem")
-    val kravhaverId: String,
-
-    @Schema(description = "Tidspunkt som det ønskes å hente gyldige perioder for")
-    val gyldigTidspunkt: LocalDateTime
 )
