+++ src/main/kotlin/no/nav/bidrag/transport/behandling/vedtak/reponse/VedtakDto.kt
@@ -2,13 +2,13 @@ package no.nav.bidrag.transport.behandling.vedtak.reponse
 
 import com.fasterxml.jackson.databind.JsonNode
 import io.swagger.v3.oas.annotations.media.Schema
-import no.nav.bidrag.behandling.felles.enums.BehandlingsrefKilde
-import no.nav.bidrag.behandling.felles.enums.EngangsbelopType
-import no.nav.bidrag.behandling.felles.enums.GrunnlagType
-import no.nav.bidrag.behandling.felles.enums.Innkreving
-import no.nav.bidrag.behandling.felles.enums.StonadType
-import no.nav.bidrag.behandling.felles.enums.VedtakKilde
-import no.nav.bidrag.behandling.felles.enums.VedtakType
+import no.nav.bidrag.domain.enums.BehandlingsrefKilde
+import no.nav.bidrag.domain.enums.EngangsbelopType
+import no.nav.bidrag.domain.enums.GrunnlagType
+import no.nav.bidrag.domain.enums.Innkreving
+import no.nav.bidrag.domain.enums.StonadType
+import no.nav.bidrag.domain.enums.VedtakKilde
+import no.nav.bidrag.domain.enums.VedtakType
 import java.math.BigDecimal
 import java.time.LocalDate
 import java.time.LocalDateTime
@@ -104,6 +104,31 @@ data class StonadsendringDto(
 )
 
 @Schema
+data class VedtakPeriodeDto(
+
+    @Schema(description = "Periode fra-og-med-dato")
+    val fomDato: LocalDate,
+
+    @Schema(description = "Periode til-dato")
+    val tilDato: LocalDate?,
+
+    @Schema(description = "Beregnet stønadsbeløp")
+    val belop: BigDecimal?,
+
+    @Schema(description = "Valutakoden tilhørende stønadsbeløpet")
+    val valutakode: String?,
+
+    @Schema(description = "Resultatkoden tilhørende  stønadsbeløpet")
+    val resultatkode: String,
+
+    @Schema(description = "Referanse - delytelsesId/beslutningslinjeId -> bidrag-regnskap. Skal fjernes senere")
+    val delytelseId: String?,
+
+    @Schema(description = "Liste over alle grunnlag som inngår i perioden")
+    val grunnlagReferanseListe: List<String>
+)
+
+@Schema
 data class EngangsbelopDto(
 
     @Schema(description = "Type Engangsbeløp. Saertilskudd, gebyr m.m.")
@@ -160,29 +185,4 @@ data class BehandlingsreferanseDto(
 
     @Schema(description = "Kildesystemets referanse til behandlingen")
     val referanse: String
-)
-
-@Schema
-data class VedtakPeriodeDto(
-
-    @Schema(description = "Periode fra-og-med-dato")
-    val fomDato: LocalDate,
-
-    @Schema(description = "Periode til-dato")
-    val tilDato: LocalDate?,
-
-    @Schema(description = "Beregnet stønadsbeløp")
-    val belop: BigDecimal?,
-
-    @Schema(description = "Valutakoden tilhørende stønadsbeløpet")
-    val valutakode: String?,
-
-    @Schema(description = "Resultatkoden tilhørende  stønadsbeløpet")
-    val resultatkode: String,
-
-    @Schema(description = "Referanse - delytelsesId/beslutningslinjeId -> bidrag-regnskap. Skal fjernes senere")
-    val delytelseId: String?,
-
-    @Schema(description = "Liste over alle grunnlag som inngår i perioden")
-    val grunnlagReferanseListe: List<String>
 )
